just a test

+++ database/migrations/2024_08_02_054132_create_aitools_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->text('description');
             $table->string('link');
             $table->boolean('hasFreePlan')->default(false);
-            $table->decimal('price')->nullable();
+            $table->integer('price')->nullable();
             $table->timestamps();
         });
     }
